Decter/requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
reportlab>=4.0.0
websockets>=11.0.0
requests>=2.31.0 